Move CLI to a separate folder "apps"
