+++ control_plane/controller/controller/apps/cli/docs/srv6pm_configuration.md
