+++ control_plane/controller/controller/apps/cli/topo_cli.py
@@ -33,7 +33,7 @@ from argparse import ArgumentParser
 
 # Controller dependencies
 from controller import arangodb_utils
-from controller.cli import utils as cli_utils
+from apps.cli import utils as cli_utils
 
 # Interval between two consecutive extractions (in seconds)
 DEFAULT_TOPO_EXTRACTION_PERIOD = 0
